Create compare function
Created the compare function in the Game class to compare each player guess to the actual code and provide feedback. It provides feedback, but code needs to be tweaked to provide accurate and useful feedback.

The decode_attempt function in the human_player.rb file was modified to provide colour options by iterating over the peg colours already stated in the assets.rb file instead of manually typing out each colour in the code.